Use same MarketTable component for Scalar and Categorical markets
