+++ app/src/MarketTable/MarketRow/marketRow.scss
@@ -50,4 +50,34 @@
       margin-right: 1rem;
     }
   }
+
+  .tab {
+    &-content {
+      line-height: normal;
+
+      h1 {
+        font-weight: 600;
+        font-size: 1.6rem;
+        margin: 2rem 0;
+      }
+      h2 {
+        font-weight: 600;
+        font-size: 1.6rem;
+        margin: 2rem 0;
+      }
+      h3 {
+        font-weight: 600;
+        font-size: 1.6rem;
+        margin: 2rem 0;
+      }
+      h4, h5, h6 {
+        font-weight: 600;
+        font-size: 1.6rem;
+      }
+
+      a {
+        color: $blue;
+      }
+    }
+  }
 }
