+++ app/src/MarketTable/deprecatedComponents/ToggleConditional.js
