+++ app/src/MarketTable/deprecatedComponents/toggleConditional.scss
