+++ app/src/MarketTable/index.js
@@ -1,21 +1,213 @@
-import React from "react";
-import CategoricalMarketTable from "./CategoricalMarketTable";
-import ScalarMarketTable from "./ScalarMarketTable";
+import React, { useEffect, useState, useMemo } from "react";
+import PropTypes from "prop-types";
+import Web3 from "web3";
+import cn from "classnames/bind";
+import Decimal from "decimal.js-light";
 
-const MarketTable = props => {
-  const { markets } = props;
-  if (markets && markets.length > 0) {
-    if (markets[0].type === "CATEGORICAL") {
-      return <CategoricalMarketTable {...props} />;
-    } else if (markets[0].type === "SCALAR") {
-      return <ScalarMarketTable {...props} />;
-    } else {
-      throw Error("Unknown market type");
+import style from "./marketTable.scss";
+
+import MarketRow from "./MarketRow";
+import Spinner from "components/Spinner";
+
+import {
+  getMarketProbabilities,
+  getStagedMarketProbabilities
+} from "utils/probabilities";
+
+const { BN } = Web3.utils;
+
+const cx = cn.bind(style);
+
+const MarketTable = ({
+  markets,
+  positions,
+  lmsrState,
+  marketSelections,
+  setMarketSelections,
+  resetMarketSelections,
+  stagedTradeAmounts,
+  openModal,
+  collateral,
+  tradeHistory
+}) => {
+  useEffect(() => {
+    resetMarketSelections();
+    // FIXME This breaks when reloading component after market maker address update
+    // return () => {
+    //   setMarketSelections(null);
+    // };
+  }, [markets]);
+
+  const [marketProbabilities, setMarketProbabilities] = useState(null);
+  const [stagedMarketProbabilities, setStagedMarketProbabilities] = useState(
+    null
+  );
+  useMemo(() => {
+    if (lmsrState != null) {
+      const { funding, positionBalances } = lmsrState;
+      const {
+        invB,
+        positionProbabilities,
+        newMarketProbabilities
+      } = getMarketProbabilities(
+        funding,
+        positionBalances,
+        markets,
+        positions,
+        marketSelections
+      );
+      setMarketProbabilities(newMarketProbabilities);
+
+      if (stagedTradeAmounts != null) {
+        const marketProbabilitiesAfterStagedTrade = getStagedMarketProbabilities(
+          {
+            positionProbabilities,
+            invB,
+            stagedTradeAmounts,
+            markets,
+            positions,
+            marketSelections
+          }
+        );
+        setStagedMarketProbabilities(marketProbabilitiesAfterStagedTrade);
+      } else {
+        setStagedMarketProbabilities(null);
+      }
     }
+  }, [lmsrState, markets, positions, marketSelections, stagedTradeAmounts]);
+
+  const conditionalDisabled = markets.length === 1;
+
+  const conditionalMarketIndices = (marketSelections || []).reduce(
+    (acc, { isAssumed }, index) => (isAssumed ? [...acc, index] : acc),
+    []
+  );
+
+  const marketsWithSetIndices = markets.map((market, index) => ({
+    ...market,
+    index
+  }));
+
+  const conditionalMarkets = [];
+  const nonConditionalMarkets = [];
+
+  marketsWithSetIndices.forEach(market => {
+    if (
+      conditionalMarketIndices.indexOf(marketsWithSetIndices.indexOf(market)) >
+      -1
+    ) {
+      conditionalMarkets.push(market);
+    } else {
+      nonConditionalMarkets.push(market);
+    }
+  });
+
+  if (!lmsrState || !marketProbabilities) {
+    return <Spinner />;
   }
 
-  // TODO return a prettier error if no markets returned
-  return null;
+  return (
+    <div className={cx("market-table")}>
+      {conditionalMarkets.map(market => (
+        <MarketRow
+          key={market.conditionId}
+          lmsrState={lmsrState}
+          tradeHistory={tradeHistory}
+          stagedProbabilities={
+            marketProbabilities != null
+              ? marketProbabilities[market.index]
+              : null
+          }
+          probabilities={
+            stagedMarketProbabilities != null
+              ? stagedMarketProbabilities[market.index]
+              : null
+          }
+          marketSelections={marketSelections}
+          setMarketSelection={setMarketSelections}
+          {...market}
+        />
+      ))}
+      {conditionalMarkets.length > 0 && (
+        <>
+          <tr className={cx("explanation-row")}>
+            <td
+              colSpan={6}
+              className={cx("explanation", "explanation-top", "arrow")}
+            >
+              <span>If</span>
+            </td>
+          </tr>
+          <tr className={cx("explanation-row")}>
+            <td
+              colSpan={6}
+              className={cx("explanation", "explanation-down", "arrow")}
+            >
+              <span>Then</span>
+            </td>
+          </tr>
+        </>
+      )}
+      {nonConditionalMarkets.map(market => (
+        <MarketRow
+          key={market.conditionId}
+          lmsrState={lmsrState}
+          collateral={collateral}
+          stagedProbabilities={
+            marketProbabilities != null
+              ? marketProbabilities[market.index]
+              : null
+          }
+          probabilities={
+            stagedMarketProbabilities != null
+              ? stagedMarketProbabilities[market.index]
+              : null
+          }
+          tradeHistory={tradeHistory}
+          disableConditional={conditionalDisabled}
+          marketSelections={marketSelections}
+          setMarketSelection={setMarketSelections}
+          {...market}
+        />
+      ))}
+    </div>
+  );
+};
+
+MarketTable.propTypes = {
+  markets: PropTypes.arrayOf(
+    PropTypes.shape({
+      conditionId: PropTypes.string.isRequired
+    }).isRequired
+  ).isRequired,
+  positions: PropTypes.arrayOf(
+    PropTypes.shape({
+      positionIndex: PropTypes.number.isRequired,
+      outcomes: PropTypes.arrayOf(
+        PropTypes.shape({
+          marketIndex: PropTypes.number.isRequired,
+          outcomeIndex: PropTypes.number.isRequired
+        }).isRequired
+      ).isRequired
+    }).isRequired
+  ).isRequired,
+  lmsrState: PropTypes.shape({
+    funding: PropTypes.instanceOf(BN).isRequired,
+    positionBalances: PropTypes.arrayOf(PropTypes.instanceOf(BN).isRequired)
+      .isRequired
+  }),
+  marketSelections: PropTypes.arrayOf(
+    PropTypes.shape({
+      selectedOutcomeIndex: PropTypes.number,
+      isAssumed: PropTypes.bool.isRequired
+    }).isRequired
+  ),
+  resetMarketSelections: PropTypes.func.isRequired,
+  setMarketSelections: PropTypes.func.isRequired,
+  stagedTradeAmounts: PropTypes.arrayOf(
+    PropTypes.instanceOf(Decimal).isRequired
+  ),
+  openModal: PropTypes.func.isRequired
 };
 
 export default MarketTable;
